Split up the feature into reasonable files
