+++ app/src/main/java/com/hedvig/app/feature/zignsec/usecase/SimpleSignStartAuthResult.kt
@@ -1,0 +1,8 @@
+package com.hedvig.app.feature.zignsec.usecase
+
+sealed class SimpleSignStartAuthResult {
+    data class Success(val url: String) :
+        SimpleSignStartAuthResult()
+
+    object Error : SimpleSignStartAuthResult()
+}
