+++ app/src/main/java/com/hedvig/app/feature/zignsec/usecase/SubscribeToAuthStatusUseCase.kt
@@ -1,0 +1,11 @@
+package com.hedvig.app.feature.zignsec.usecase
+
+import com.apollographql.apollo.ApolloClient
+import com.apollographql.apollo.coroutines.toFlow
+import com.hedvig.android.owldroid.graphql.AuthStatusSubscription
+
+class SubscribeToAuthStatusUseCase(
+    private val apolloClient: ApolloClient,
+) {
+    operator fun invoke() = apolloClient.subscribe(AuthStatusSubscription()).toFlow()
+}
