+++ app/src/main/java/com/hedvig/app/util/extensions/FragmentExtensions.kt
@@ -1,6 +1,7 @@
 package com.hedvig.app.util.extensions
 
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.FragmentTransaction
 import androidx.lifecycle.coroutineScope
 
 val Fragment.screenWidth: Int
@@ -8,3 +9,5 @@ val Fragment.screenWidth: Int
 
 val Fragment.viewLifecycleScope
     get() = viewLifecycleOwner.lifecycle.coroutineScope
+
+fun FragmentTransaction.addToBackStack() = addToBackStack(null)
